+++ trpm-web/src/main/java/com/vipkid/portal/classroom/model/PeSupervisorCommentsVo.java
@@ -1,20 +1,20 @@
 package com.vipkid.portal.classroom.model;
 
-import java.util.List;
-import java.util.Map;
-
 import com.vipkid.rest.validation.annotation.Ignore;
 import com.vipkid.rest.validation.annotation.Length;
 import com.vipkid.rest.validation.annotation.NotNull;
 
+import java.util.List;
+import java.util.Map;
+
 @NotNull
 public class PeSupervisorCommentsVo {
 
-	private Long id;
+	private Integer id;
 	
 	private Integer peId;
 	
-	private Long onlineClassId;
+	private Integer onlineClassId;
 	
 	private Integer delayDays;
 	
@@ -68,11 +68,11 @@ public class PeSupervisorCommentsVo {
 	
 	private String submitType;
 	
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public Long getOnlineClassId() {
+	public Integer getOnlineClassId() {
 		return onlineClassId;
 	}
 
@@ -156,7 +156,7 @@ public class PeSupervisorCommentsVo {
 		return result;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 	
@@ -168,7 +168,7 @@ public class PeSupervisorCommentsVo {
 		this.peId = peId;
 	}
 
-	public void setOnlineClassId(Long onlineClassId) {
+	public void setOnlineClassId(Integer onlineClassId) {
 		this.onlineClassId = onlineClassId;
 	}
